requirements.txt
google-genai>=0.8.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
moviepy>=1.0.3
aiohttp>=3.8.0
torch>=2.0.0
tqdm>=4.65.0